[refactor] Improve modal structure
Modal component 개수 증가로 구조 변경
- 모든 모달 컴포넌트는 Modals에 속함
- 기존 Modal은 ModalFrame으로 변경

refactor #3
