+++ src/components/Modals/ModalFrame/ModalFrame.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import styled from '../../lib/styled';
+import styled from '../../../lib/styled';
 import { ModalType, MODAL_TYPE_DEAFULT } from './index';
-import Portal from '../../lib/Portal';
+import Portal from '../../../lib/Portal';
 // import { createPortal } from 'react-dom';
 
 /* 
@@ -71,7 +71,7 @@ const ModalInner = styled.div<WrapperProps>`
   padding: 30px 10px;
 `;
 
-const Modal = ({ width, visible, children, onClose, closable, type }: Props): JSX.Element => {
+const ModalFrame = ({ width, visible, children, onClose, closable, type }: Props): JSX.Element => {
   const onMaskClick = (e: React.MouseEvent) => {
     if (e.target === e.currentTarget) {
       onClose(e);
@@ -99,4 +99,4 @@ const Modal = ({ width, visible, children, onClose, closable, type }: Props): JS
   );
 };
 
-export default Modal;
+export default ModalFrame;
